cleanup: changed stack structure of gameslots page, added header w/ sidemenu everywhere, corrected 1 alert dialog, standardised back button
